+++ addon/components/table-of-contents-plugin/toolbar-button.ts
@@ -38,12 +38,43 @@ export default class TableOfContentsCardComponent extends Component<Args> {
       );
     } else {
       const { schema } = this.controller;
-      this.controller.withTransaction(
-        (tr) => {
-          return tr.replaceRangeWith(0, 0, schema.node('table_of_contents'));
-        },
-        { view: this.controller.mainEditorView }
-      );
+      const state = this.controller.activeEditorState;
+      let replacePosition: number | undefined = undefined;
+      state.doc.descendants((node, pos, parent, index) => {
+        if (
+          replacePosition === undefined &&
+          state.doc.canReplaceWith(
+            index,
+            index,
+            schema.nodes['table_of_contents']
+          )
+        ) {
+          replacePosition = pos;
+        } else if (
+          index === state.doc.childCount - 1 &&
+          replacePosition === undefined &&
+          state.doc.canReplaceWith(
+            index + 1,
+            index + 1,
+            schema.nodes['table_of_contents']
+          )
+        ) {
+          replacePosition = pos + node.nodeSize;
+        }
+        return false;
+      });
+      if (replacePosition !== undefined) {
+        this.controller.withTransaction(
+          (transaction) => {
+            return transaction.replaceWith(
+              replacePosition as number,
+              replacePosition as number,
+              schema.node('table_of_contents')
+            );
+          },
+          { view: this.controller.mainEditorView }
+        );
+      }
     }
   }
 }